Rename utils to util for consistency (look at parent directory).
Also moved some visualization code from util into visualize direct.
